refactor(linebreak): lift the prefix sums and DOM geometry out
linebreaker.ts and breaker.ts were 844 and 823 lines. Moved the parts
that are separate concerns rather than slicing the algorithms:

  dom/geometry.ts  element reads and the post-write layout check
  policy.ts        engineLimits, beside the defaults it resolves
  layout/sums.ts   prefix-sum construction, one interface: items+flex

linebreaker.ts 844 -> 799, breaker.ts 823 -> 745.

sums.ts crosses a module boundary on the first-break path, so it was
measured rather than assumed. perf/time.ts, 7 reps, 21 cells: warm 474.3
ms with the split against 475.8 without, and two runs of identical code
differ by 0.8%. No cost outside noise. All three digests reproduce and
466 unit tests pass, so the layout is unchanged.
