+++ packages/linebreak/src/policy.ts
@@ -1,3 +1,5 @@
+import type { LinebreakerOptions } from "./types"
+
 export const engineDefaults = Object.freeze({
   minimumWidth: 240,
   safetyMargin: 0.5,
@@ -8,4 +10,12 @@ export const engineDefaults = Object.freeze({
   expansionBudget: 0.02,
   trackingBudget: 0.03,
   lastLineMinWidth: 0,
+})
+
+export const engineLimits = (options: LinebreakerOptions) => ({
+  minimumWidth: options.minimumWidth ?? engineDefaults.minimumWidth,
+  safetyMargin: options.safetyMargin ?? engineDefaults.safetyMargin,
+  maximumRetries: options.retries ?? engineDefaults.retries,
+  maximumCharacters:
+    options.maximumCharacters ?? engineDefaults.maximumCharacters,
 })
